+++ GameData/RedOnion/Scripts/printmodules.lua
@@ -1,6 +1,0 @@
-for part in ship.native.parts do
-    for module in part.modules do
-        print(module)
-    end
-end
-