+++ RedOnion.KSP/MunOS/Process.cs
@@ -157,7 +157,6 @@ namespace RedOnion.KSP.MunOS
 			public ShutdownHook(T target)
 			{
 				_target = new WeakReference<T>(target);
-				Value.DebugLog("ShutdownHook created for process #{0}", currentId);
 				process = current;
 				process.shutdown += Shutdown;
 			}
@@ -171,7 +170,6 @@ namespace RedOnion.KSP.MunOS
 			{
 				if (process == null)
 					return;
-				Value.DebugLog("ShutdownHook disposed for process #{0}", process.id);
 				process.shutdown -= Shutdown;
 				process = null;
 				_target = null;