+++ RedOnion.ROS/Parsing/Functions.cs
@@ -1,9 +1,7 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Diagnostics;
-using RedOnion.ROS.Utilities;
+using RedOnion.Collections;
 
 namespace RedOnion.ROS.Parsing
 {